requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
blinker==1.6.3
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
distro==1.8.0
env==0.1.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==1.3.7
packaging==23.2
pluggy==1.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.22
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.0
